Move the go to action to the leaf more menu
